Reorginized file structure and made necessary path changes
