+++ frontend/index.html
@@ -5,10 +5,10 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Document</title>
-        <link rel="stylesheet" type="text/css" href="./css/styles.css" />
+        <link rel="stylesheet" type="text/css" href="/css/styles.css" />
     </head>
     <body>
         <div id="app-container"></div>
     </body>
-    <script src="./bundle/bundle.js"></script> 
+    <script src="/bundle/bundle.js"></script> 
 </html>
