+++ package.json
@@ -1,10 +1,12 @@
 {
   "scripts": {
-    "dev": "nodemon --watch src/js server.js"
+    "dev": "nodemon --watch frontend/js backend/server.js"
   },
   "dependencies": {
     "express": "^4.18.2",
-    "phaser": "^3.55.2"
+    "path": "^0.12.7",
+    "phaser": "^3.55.2",
+    "sqlite3": "^5.1.6"
   },
   "devDependencies": {
     "nodemon": "^2.0.21",
